python_workers/easynet_worker/requirements.txt
redis
requests
python-dotenv
PyYAML
